+++ NEWS.md
@@ -1,5 +1,7 @@
 # strzip 0.0.1.9000
 
+* Read/Write compressed files
+
 # strzip 0.0.1
 
 * Integrated SMAZ